requirements.txt
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
torch==2.6.0
tqdm==4.67.1
transformers==4.45.2
